Move tests to top level
This makes autodiscovery work with `python -m unittest`
