+++ tox.ini
@@ -6,29 +6,7 @@ skipsdist=True
 [testenv]
 deps = -rrequirements_test.txt
 commands =
-	python -m luna.gateware.interface.spi
-	python -m luna.gateware.interface.ulpi
-	python -m luna.gateware.interface.psram
-	python -m luna.gateware.interface.uart
-	python -m luna.gateware.architecture.car
-	python -m luna.gateware.utils.cdc
-	python -m luna.gateware.debug.ila
-	python -m luna.gateware.stream.generator
-	python -m luna.gateware.usb.stream
-	python -m luna.gateware.usb.usb2.packet
-	python -m luna.gateware.usb.usb2.request
-	python -m luna.gateware.usb.usb2.control
-	python -m luna.gateware.usb.usb2.device
-	python -m luna.gateware.usb.usb2.transfer
-	python -m luna.gateware.usb.usb2.descriptor
-	python -m luna.gateware.usb.usb3.physical.scrambling
-	python -m luna.gateware.usb.usb3.physical.ctc
-	python -m luna.gateware.usb.usb3.physical.lfps
-	python -m luna.gateware.usb.usb3.link.receiver
-	python -m luna.gateware.usb.usb3.link.data
-	python -m luna.gateware.usb.usb3.link.crc
-	python -m luna.gateware.usb.usb3.application.request
-	python -m luna.gateware.memory
+	python -m unittest discover -t . -s tests/
 
 [gh-actions]
 python =
